+++ packages/fcl/src/config-utils.js
@@ -44,13 +44,5 @@ export async function buildAuthnConfig() {
 
   const appDomainTag = await config.get("fcl.appDomainTag")
 
-  invariant(
-    discoveryWallet != null,
-    `
-      Required value for discovery.wallet not defined in config.
-      See: "https://github.com/onflow/flow-js-sdk/blob/master/packages/fcl/src/exec/query.md#configuration"
-    `
-  )
-
   return {discoveryWallet, discoveryWalletMethod, appDomainTag}
 }