+++ packages/fcl/src/current-user/index.js
@@ -93,7 +93,7 @@ function notExpired(user) {
   )
 }
 
-async function authenticate(opts = {redir: false}) {
+async function authenticate({ service, redir = false }) {
   return new Promise(async (resolve, reject) => {
     spawnCurrentUser()
     const user = await snapshot()
@@ -102,8 +102,16 @@ async function authenticate(opts = {redir: false}) {
     const {discoveryWallet, discoveryWalletMethod, appDomainTag} =
       await buildAuthnConfig()
 
+    invariant(
+      service || discoveryWallet,
+      `
+        If no service passed to "authenticate," then "discovery.wallet" must be defined in config.
+        See: "https://docs.onflow.org/fcl/reference/api/#setting-configuration-values"
+      `
+    )
+
     const suppressRedirWarning = await config.get("fcl.warning.suppress.redir")
-    if (opts.redir && !suppressRedirWarning) {
+    if (redir && !suppressRedirWarning) {
       console.warn(
         `You are manually enabling a very experimental feature that is not yet standard, use at your own risk.
          You can disable this warning by setting fcl.warning.suppress.redir to true in your config`
@@ -112,7 +120,7 @@ async function authenticate(opts = {redir: false}) {
 
     try {
       const response = await execService({
-        service: {
+        service: service || {
           type: "authn",
           endpoint: discoveryWallet,
           method: discoveryWalletMethod,
@@ -122,7 +130,7 @@ async function authenticate(opts = {redir: false}) {
           appDomainTag,
           extensions: window.fcl_extensions || [],
         },
-        opts,
+        opts: { redir },
       })
       send(NAME, SET_CURRENT_USER, await buildUser(response))
     } catch (e) {