+++ packages/fcl/src/discovery/index.js
@@ -1,0 +1,7 @@
+import authn from "./services/authn"
+
+const discovery = {
+  authn
+}
+
+export {discovery}